Removed wrappring of typedefs and restored report file
o When a type defined with typedef or using was need for a method wrapper, a Julia wrapper was generated for this type. This was causing problem if a wrapper was also generated for the underlying type or another typedef of it. With this change, it is the underlying type (after resolving all typedefs/usings), which is wrapped, avoiding this way multiple type for the same underlying type.

o A report file is now generated. For time being it contains the list of types needed to generate the wrapper but whose definition was not found.