Conditional finalizer generation and improved constructor wrapper generation
o With this change the finalizer that deletes an object allocated
from Julia is not generated if the class destructor is not public
or is deleted. The generation can also be disabled for each class
by configuration (parameter vetoed_finalizer_classes)

o The tool is now better in deciding when to generate a wrapper for
the class default constructors.