Improve CMakeLists.txt
This is a large change to CMakeLists, with the headline that it fixes
the compilation of wrapit on AlmaLinux9.

Simplify the configuration to use the system clang/llvm installation
(use of CLANG_JLL from the Julia installation seems very fragile and I
could not get it to work properly)

Use the "unified" new clang targets (libclang clang-cpp) instead of the
deprecated split libraries.

Use recommended llvm recipe for the code llvm libraries.

Update CMake minimum version to 3.12, which throws no warnings on
configuration with the latest releases (released 5 years ago, so not
onerous!).

Update README with instructions for Alma and updated instructions for
OS X brew install location.