Improve wrapper declaration ordering in order to reduce "missing factory" error occurrences and more
1. wrapper declaration order.

Caveat: CxxWrap requires the wrapper of a type to be declared before
the wrappers of types or methods that depend on it.

Before this change, the required ordering was guaranteed for type-to-type
dependencies coming from inheritance/supertyping and method-to-type of
non-template class methods.

The change extends it to cover type-to-type dependencies coming from template
parameters and method-to-type dependency of template class methods.

The new ordering strategy is:

1. wrappers of parent classes declared before the ones of their children;
2. wrappers of template parameters of a class declared before its
own wrapper;
3. wrappers of methods of non-template class declared after all
wrappers of all types;
4. wrappers of methods of template class declared at the same time
as the class wrapper and after the wrapper of all types needed by
them.

When such ordering does not exist, a message is reported on stderr
and in the report file.

The test TestOrder has been added to test the featur.

2. improved std::vector and std::valarray support

Types used as a class template parameter are now autowrap it autowrap
is enabled.