Work on cmake build system, unit tests and ex001 example.
1. cmake build.

Add support for projects with wrapit generating several generated files and
better dependency handling.

wrapit is now run at cmake configuration step. It handles dependencies:
wrapit rerun if the .wit(.in) configuration file or a header file the
code produced by wrapit depends on is touched.

2. runtests.jl

Fixed result stacking issue in the test summary when running test in spawned
processes and move to this spawned-process scheme.

2. ex001 example

Makefile of ex001-HelloWorld was updated to behace the same as with the
cmake-based build. We will keep both Makefile and CMakeLists.txt for
this example in order to provide an example with both systems.

build.jl of ex001-HelloWorld is now also installing the generated wrapper
module. The script was renamed install.jl.

Next step:

Create a cmake function that will take a .wit file as inputs and generate the
directive to generate and build the library. A --get-config option will be added
to wrapit, in order to acces to the configuration stored in the .wit file from
cmake. This function can then be used for unit test, examples and users' own
code.